Updated Quiz naming and App routes
