+++ src/App.tsx
@@ -17,7 +17,9 @@ import Signup from './components/Signup.tsx';
 import Flashcard from './components/Flashcards/Flashcard.tsx';
 import Tutorial from './pages/Tutorial.tsx';
 import AttackQuiz from './components/attacks_quiz/AttackQuiz.tsx';
+import Quiz from './components/Quizzes/Quiz.tsx';
 import * as FlashcardData from './components/Flashcards/FlashcardData';
+import LevelRSA from './components/attacks_quiz/LevelRSA.tsx';
 
 const baseStyle = {
   indicator: {
@@ -120,7 +122,6 @@ const App = () => {
         <Route path="level">
           <Route path="quiz1" element={<AttackQuiz weekNumber={1} taskID={'quiz1'} />} />
           <Route path="task1" element={<CCEmail weekNumber={1} taskID={'task1'} />} />
-          {/*<Route path="3" element={<LevelRSA />} />*/}
           <Route
             path="flashcards1-attacks"
             element={<Flashcard content={FlashcardData.week1_attacks} weekNumber={1} taskID={'flashcards1-attacks'} />}
@@ -134,6 +135,16 @@ const App = () => {
             path="flashcards2-crypto"
             element={<Flashcard content={FlashcardData.week2_crypto} weekNumber={2} taskID={'flashcards2-crypto'} />}
           />
+          <Route
+            path="flashcards2-packets"
+            element={<Flashcard content={FlashcardData.week2_packets} weekNumber={2} taskID={'flashcards2-packets'} />}
+          />
+          <Route
+            path="flashcards2-integrity"
+            element={<Flashcard content={FlashcardData.week2_message_integrity} weekNumber={2} taskID={'flashcards2-integrity'} />}
+          />
+          <Route path="quiz2" element={<Quiz weekNumber={2} taskID={'quiz2'} />} />
+          <Route path="rsa" element={<LevelRSA />} />
         </Route>
         <Route path="sandbox" element={<Sandbox />} />
         <Route
